first step of splitting
