+++ plugin/commonInterfaces.ts
@@ -1,0 +1,9 @@
+export interface BabelMapType {
+  version: number;
+  sources: string[];
+  names: string[];
+  sourceRoot?: string | undefined;
+  sourcesContent?: string[] | undefined;
+  mappings: string;
+  file: string;
+}
