+++ plugin/commonObjects.ts
@@ -1,0 +1,115 @@
+/**
+ * holds a map of function names as keys and array of argument indexes as values which should be automatically workletized(they have to be functions)(starting from 0)
+ */
+export const functionArgsToWorkletize = new Map([
+  ['useFrameCallback', [0]],
+  ['useAnimatedStyle', [0]],
+  ['useAnimatedProps', [0]],
+  ['createAnimatedPropAdapter', [0]],
+  ['useDerivedValue', [0]],
+  ['useAnimatedScrollHandler', [0]],
+  ['useAnimatedReaction', [0, 1]],
+  ['useWorkletCallback', [0]],
+  // animations' callbacks
+  ['withTiming', [2]],
+  ['withSpring', [2]],
+  ['withDecay', [1]],
+  ['withRepeat', [3]],
+]);
+
+export const objectHooks = new Set([
+  'useAnimatedGestureHandler',
+  'useAnimatedScrollHandler',
+]);
+
+export const globals = new Set([
+  'this',
+  'console',
+  'performance',
+  '_chronoNow',
+  'Date',
+  'Array',
+  'ArrayBuffer',
+  'Int8Array',
+  'Int16Array',
+  'Int32Array',
+  'Uint8Array',
+  'Uint8ClampedArray',
+  'Uint16Array',
+  'Uint32Array',
+  'Float32Array',
+  'Float64Array',
+  'Date',
+  'HermesInternal',
+  'JSON',
+  'Math',
+  'Number',
+  'Object',
+  'String',
+  'Symbol',
+  'undefined',
+  'null',
+  'UIManager',
+  'requestAnimationFrame',
+  '_WORKLET',
+  'arguments',
+  'Boolean',
+  'parseInt',
+  'parseFloat',
+  'Map',
+  'WeakMap',
+  'WeakRef',
+  'Set',
+  '_log',
+  '_scheduleOnJS',
+  '_makeShareableClone',
+  '_updateDataSynchronously',
+  'eval',
+  '_updatePropsPaper',
+  '_updatePropsFabric',
+  '_removeShadowNodeFromRegistry',
+  'RegExp',
+  'Error',
+  'ErrorUtils',
+  'global',
+  '_measure',
+  '_scrollTo',
+  '_dispatchCommand',
+  '_setGestureState',
+  '_getCurrentTime',
+  '_eventTimestamp',
+  '_frameTimestamp',
+  'isNaN',
+  'LayoutAnimationRepository',
+  '_notifyAboutProgress',
+  '_notifyAboutEnd',
+]);
+
+export const gestureHandlerGestureObjects = new Set([
+  // from https://github.com/software-mansion/react-native-gesture-handler/blob/new-api/src/handlers/gestures/gestureObjects.ts
+  'Tap',
+  'Pan',
+  'Pinch',
+  'Rotation',
+  'Fling',
+  'LongPress',
+  'ForceTouch',
+  'Native',
+  'Manual',
+  'Race',
+  'Simultaneous',
+  'Exclusive',
+]);
+
+export const gestureHandlerBuilderMethods = new Set([
+  'onBegin',
+  'onStart',
+  'onEnd',
+  'onFinalize',
+  'onUpdate',
+  'onChange',
+  'onTouchesDown',
+  'onTouchesMove',
+  'onTouchesUp',
+  'onTouchesCancelled',
+]);
